# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Quản Lý Cân Xe API",
    "version": "1.0.0",
    "description": "API tài liệu cho hệ thống quản lý cân xe.\n\nHệ thống này cung cấp API để quản lý việc cân xe, thông tin khách hàng, \nhàng hóa và các chức năng liên quan đến quản lý phiếu cân.\n",
    "license": {
      "name": "Licensed Under MIT",
      "url": "https://spdx.org/licenses/MIT.html"
    },
    "contact": {
      "name": "Your Name",
      "url": "https://yourwebsite.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Development server"
    },
    {
      "url": "https://api.example.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Form": {
        "type": "object",
        "properties": {
          "formId": {
            "type": "number",
            "description": "ID form"
          },
          "ten": {
            "type": "string",
            "description": "Tên form"
          },
          "form": {
            "type": "string",
            "description": "Tên file form"
          },
          "menuname": {
            "type": "string",
            "description": "Tên menu"
          },
          "vitri": {
            "type": "string",
            "description": "Vị trí menu"
          }
        },
        "example": {
          "formId": 1,
          "ten": "Thông số hệ thống",
          "form": "frmCaidat",
          "menuname": "mnuThongsohethong",
          "vitri": "danhmuc"
        }
      },
...
      "NhomQuyen": {
        "type": "object",
        "required": [
          "ma",
          "ten"
        ],
        "properties": {
          "nhomId": {
            "type": "number",
            "description": "ID tự động tạo của nhóm quyền"
          },
          "ma": {
            "type": "string",
            "description": "Mã nhóm quyền"
          },
          "ten": {
            "type": "string",
            "description": "Tên nhóm quyền"
          }
        },
        "example": {
          "nhomId": 1,
          "ma": "admin",
          "ten": "Admin"
        }
      },
...
      "Quyen": {
        "type": "object",
        "required": [
          "nhomId",
          "formId"
        ],
        "properties": {
          "quyenId": {
            "type": "number",
            "description": "ID tự động tạo của quyền"
          },
          "nhomId": {
            "type": "number",
            "description": "ID nhóm quyền"
          },
          "formId": {
            "type": "number",
            "description": "ID form"
          },
          "nhomQuyen": {
            "$ref": "#/components/schemas/NhomQuyen"
          },
          "form": {
            "$ref": "#/components/schemas/Form"
          }
        },
        "example": {
          "quyenId": 1,
          "nhomId": 1,
          "formId": 1,
          "form": {
            "formId": 1,
            "ten": "Thông số hệ thống"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/nhomquyen/{nhomId}/permissions/{formId}": {
      "post": {
        "tags": [
          "Nhóm Quyền"
        ],
        "summary": "Thêm quyền vào nhóm quyền",
        "description": "Thêm một quyền (form) vào nhóm quyền",
        "operationId": "addPermissionToGroup",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "nhomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "ID nhóm quyền"
          },
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "ID form (quyền)"
          }
        ],
        "responses": {
          "200": {
            "description": "Thêm quyền thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Thêm quyền thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Quyen"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Không có quyền truy cập (Chỉ admin)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Không có quyền truy cập"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "tags": [
          "Nhóm Quyền"
        ],
        "summary": "Xóa quyền khỏi nhóm quyền",
        "description": "Xóa một quyền (form) khỏi nhóm quyền",
        "operationId": "removePermissionFromGroup",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "nhomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "ID nhóm quyền"
          },
          {
            "name": "formId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "ID form (quyền)"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa quyền thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Xóa quyền thành công"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Không có quyền truy cập (Chỉ admin)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Không có quyền truy cập"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/nhomquyen/{nhomId}/permissions": {
      "put": {
        "tags": [
          "Nhóm Quyền"
        ],
        "summary": "Cập nhật danh sách quyền của nhóm",
        "description": "Cập nhật toàn bộ danh sách quyền của một nhóm quyền",
        "operationId": "updateGroupPermissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "nhomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "number"
            },
            "description": "ID nhóm quyền"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "formIds"
                ],
                "properties": {
                  "formIds": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "description": "Danh sách ID của các form (quyền)"
                  }
                },
                "example": {
                  "formIds": [
                    1,
                    2,
                    3,
                    5,
                    8
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật quyền thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Cập nhật quyền thành công"
                    },
                    "data": {
                      "type": "null"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "403": {
            "description": "Không có quyền truy cập (Chỉ admin)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "message": {
                      "type": "string",
                      "example": "Không có quyền truy cập"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFoundError"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  }
}